+++ CHANGELOG.md
@@ -588,3 +588,4 @@ All notable changes to this project will be documented in this file.
 - remove usage of boolean type in tooltips and signatures to avoid confusion since technically booleans do not exist
 - minor improvement to index expression type analyzing in UI
 - use modified transformer to get namespace for type in UI
+- fix type resolve within slice expression in UI