+++ CHANGELOG.md
@@ -589,3 +589,5 @@ All notable changes to this project will be documented in this file.
 - minor improvement to index expression type analyzing in UI
 - use modified transformer to get namespace for type in UI
 - fix type resolve within slice expression in UI
+- improve type resolve for assignments using locals, globals or outer prefix in UI
+- extend namespace find method to use assignment instead of namespaces coming from parser in UI