+++ CHANGELOG.md
@@ -593,3 +593,4 @@ All notable changes to this project will be documented in this file.
 - extend namespace find method to use assignment instead of namespaces coming from parser in UI
 - fix possibly wrong start position of member, index and call expression in UI
 - inject map constructor namespaces in UI
+- inject list constructor namespaces in UI